[FLINK-9386] Embed netty router
This commit replaces netty-router dependency with our own version of it, which is
simplified and adds guarantees about order of matching router patterns.

This is a prerequisite for FLINK-3952. netty-router 1.10 is incompatible with
Netty 4.1, while netty-router 2.2.0 brakes a compatibility in a way that we
were unable to use it.

This closes #6031.