[FLINK-16033][table-api] Added entry point in Table API Java DSL for creating intervals
This commit introduces methods for creating intervals from Java DSL table API. The two introduced methods should subsume scala equivalents of .millis, .hours, .years, .months etc.

The literals are limited to the resolution that is supported by the runtime. That means for YEAR MONTH interval the total number of months for the literal must fit into int. The created literal will be of type MONTH. For DAY SECOND intervals it supports interval of millisecond precision. The number of milliseconds for interval must fit into long. The produced literal will be of type SECOND(3).