[FLINK-16913] Migrate StateBackends to use ReadableConfig instead of Configuration
StateBackendFactories do not need a full read and write access to the
Configuration object. It's sufficient to have read only access. Moreover
the ReadableConfig is a lightweight interface that can be implemented in
other ways, not just through the Configuration. Lastly we exposed this
lightweight interface as a configuration entry point for
ExecutionEnvironments. This change will make it possible to pass the
ReadableConfig directly to the StateBackendFactories without fragile
adapters.